+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 [Discussions about unicorn-binance-websocket-api releases!](https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/discussions/categories/releases)
 
 ## 1.43.2.dev (development stage/unreleased/unstable)
+### Fixed
+- Restore backward compatibility Python >= 3.7 [issue#311](https://github.com/LUCIT-Systems-and-Development/unicorn-binance-websocket-api/issues/311)
 
 ## 1.43.2
 ### Changed